allow users to prevent overwrite of local jboss config from repository
when setting the OPENSHIFT_JBOSSEAP_CFG_LOCAL_EDIT variable to true,
the standalone.xml will not be overwritten by repository content
on startup.  this allows users to edit the config locally in the gear
via jboss-cli, the jboss console, or by hand
