+++ cartridges/openshift-origin-cartridge-jbossas/metadata/managed_files.yml
@@ -14,6 +14,7 @@ locked_files:
 - env/OPENSHIFT_JBOSSAS_JDK*
 - env/OPENSHIFT_JBOSSAS_VERSION
 - env/M2_HOME
+- jboss_cfg_backup/
 processed_templates:
 - 'metadata/jenkins_artifacts_glob.erb'
 dependency_dirs:
