+++ cartridges/openshift-origin-cartridge-jbossas/versions/7/bin/standalone.conf
@@ -15,12 +15,24 @@ function print_sed_exp_replace_env_var {
 
 CART_DIR=${OPENSHIFT_JBOSSAS_DIR}
 
+# if the repo contains a standalone.xml, back it up
+if [ -f ${OPENSHIFT_REPO_DIR}.openshift/config/standalone.xml ]
+then
+  cp ${OPENSHIFT_REPO_DIR}.openshift/config/standalone.xml $OPENSHIFT_JBOSSAS_DIR/jboss_cfg_backup/standalone.xml || :
+fi
+
 # Create a link for each file in user config to server standalone/config
 if [ -d ${OPENSHIFT_REPO_DIR}.openshift/config ]
 then
   for f in ${OPENSHIFT_REPO_DIR}.openshift/config/*
   do
     target=$(basename $f)
+    if [ $target = "standalone.xml" ]
+    then
+      # don't replace standalone.xml from the repository, we'll make a decision
+      # about whether to do that or not later.
+      continue
+    fi  
     # Remove any target that is being overwritten
     if [ -e "${CART_DIR}/standalone/configuration/$target" ]
     then
@@ -31,6 +43,13 @@ then
   done
 fi
 
+# Whether or not we found a standalone.xml in the repo, copy the last version of the standalone.xml
+# that came from the repo, to be the config, unless the user is in jbosscli mode in which case
+# we don't touch standalone.xml.
+if [ -f $OPENSHIFT_JBOSSAS_DIR/jboss_cfg_backup/standalone.xml ] && [ x$DISABLE_OPENSHIFT_MANAGED_SERVER_CONFIG != "xtrue" ] 
+then
+  cp $OPENSHIFT_JBOSSAS_DIR/jboss_cfg_backup/standalone.xml ${OPENSHIFT_JBOSSAS_DIR}/standalone/configuration/standalone.xml || :
+fi
 
 MYSQL_ENABLED="false"
 if [ -n "$OPENSHIFT_MYSQL_DB_URL" ]
