+++ cartridges/openshift-origin-cartridge-jbosseap/bin/setup
@@ -14,3 +14,6 @@ cp ${OPENSHIFT_JBOSSEAP_DIR}/versions/${version}/bin/* ${OPENSHIFT_JBOSSEAP_DIR}
 
 # Initialize the JBoss CLI history file
 [ -f ${OPENSHIFT_HOMEDIR}/.jboss-cli-history ] || touch ${OPENSHIFT_HOMEDIR}/.jboss-cli-history
+
+mkdir -p $OPENSHIFT_JBOSSEAP_DIR/jboss_cfg_backup
+cp -n ${OPENSHIFT_JBOSSEAP_DIR}/versions/${version}/standalone/configuration/standalone.xml $OPENSHIFT_JBOSSEAP_DIR/jboss_cfg_backup
