+++ cartridges/openshift-origin-cartridge-jbosseap/bin/upgrade
@@ -11,3 +11,13 @@ new_cart_version="$3"
 if [ -e $OPENSHIFT_HOMEDIR/jbosseap/metadata/jenkins_shell_command ]; then
   rm -f $OPENSHIFT_HOMEDIR/jbosseap/metadata/jenkins_shell_command
 fi
+
+# For migration, if the user repo does not contain a standalone.xml, populate the "backup" from
+# the current config file used by the cartridge.
+mkdir -p $OPENSHIFT_JBOSSEAP_DIR/jboss_cfg_backup
+if [ ! -f ${OPENSHIFT_REPO_DIR}/.openshift/config/standalone.xml ]
+then
+  cp -n ${OPENSHIFT_JBOSSEAP_DIR}/standalone/configuration/standalone.xml $OPENSHIFT_JBOSSEAP_DIR/jboss_cfg_backup
+else
+  cp -n ${OPENSHIFT_REPO_DIR}/.openshift/config/standalone.xml $OPENSHIFT_JBOSSEAP_DIR/jboss_cfg_backup
+fi
