+++ cartridges/openshift-origin-cartridge-jbosseap/metadata/managed_files.yml
@@ -13,6 +13,7 @@ locked_files:
 - env/OPENSHIFT_JBOSSEAP_LOG_DIR
 - env/OPENSHIFT_JBOSSEAP_JDK*
 - env/M2_HOME
+- jboss_cfg_backup/
 processed_templates:
 - 'metadata/jenkins_artifacts_glob.erb'
 setup_rewritten:
